+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * (Add) PrusaSlicer Printer: Nova Bene4 Mono
 * (Add) CWS: Support the GRAY2RGB and RBG2GRAY encoding for Bene Mono
 * (Add) Layer issue Z map paired with layer navigation tracker bar
+* (Improvement) When zoom into issue or drawing now it checks bounds of zoom rectangle and only performs ZoomToFit is it will be larger then the viewPort after zoom. Otherwise, it will zoom to the fixed zoom level (Auto zoom to region setting dropped as merged into this) (#42)
 * (Change) Shortcuts: + and - to go up and down on layers were change to W and S keys. Reason: + and - are bound to zoom and can lead to problems
 * (Upgrade) OpenCV from 4.2 to 4.3
 * (Remove) Menu - Tools - Layer Removal and Layer clone for redudancy they now home at layer preview toolbar under "Actions" dropdown button